Create Strings.de.resx
German translation. 
unfortunately untested (as I have no clue how to test it on my local installation)
As the translation was undertaken by using an resx editor > the file format itself should be fine.
While working on the translation, I looked into the DS4Windows GUI to see, where the original strings were used and which translation would make most sense.
So it should be fine.